requirements.txt
numpy
scipy
tensorflow
matplotlib
opencv-python-headless
pillow
streamlit

